fixed bug encountered with USA matrix test
When using cached edge expansion with the CH edges, the Path4CH object
which expanded a single edge wasn't being given the correct reverseOrder
state before from the PathBiDirRef object extracting the distances and
times from the path. This led to trying to read invalid / empty data
from a contracted edge in some circumstances, as the data wasn't
available in the correct direction.