bug fix with exported excel
The exported excel was in a state where it thought the hidden odl studio 'schema' sheet was selected. This prevented excel from doing various actions, e.g. creating a new sheet. The word 'group' also appeared at the top of the screen after the sheet or filename. We added a couple of lines to ensure the schema sheet is not selected